requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0.5
django-import-export==2.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.1.0
idna==2.9
importlib-metadata==1.6.0
MarkupPy==1.14
more-itertools==8.2.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
PyYAML==5.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==3.0.0
urllib3==1.25.9
wcwidth==0.1.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.1.0
django-filter==2.4.0
